User-Defined Function (UDF) object

User-Defined Function (UDF) object

User-Defined Function (UDF) object

Details

UDF objects store the definition of a user-defined function along with metadata and unit tests.

Methods

Public methods

udf$new()

udf$write()

udf$create()

udf$drop()

udf$test()

udf$clone()

Method new()

Initialize a md object

Usage

udf$new(
  name,
  version,
  returns,
  volatility,
  language,
  params = list(),
  body,
  description,
  schema = NA,
  replace = FALSE,
  authors = list(),
  tests = list(),
  ...
)

Arguments

name

The UDF's name (e.g., "f_calculate_stuff")

version

A version string (e.g., "1.0")

returns

Data type returned by this UDF (e.g., "float")

volatility

Given the same inputs, the function will always return the same result ("immutable"), will return the same result during a given statement ("stable"), may return different results on successive calls ("volatile").

language

String indicating the language used for the UDF's body (e.g., "sql")

params

A list of zero or more udfparameter objects specifying input parameters

body

The body of the UDF containing SQL, Python, or other supported language

description

A detailed description of the UDF

schema

Database schema in which to place UDF (defualt: NA)

replace

Logical value indicating whether existing versions should be replaced (default: FALSE)

authors

List of authors (e.g., list("Jane Doe"))

tests

A list of zero or more udftest objects specifying test cases for this UDF

Method write()

Write UDF definition file

Usage

udf$write(file, ...)

Arguments

...

Additional arguments passed to yaml::write_yaml()

File

File to write to

Method create()

Usage

udf$create(conn, test = TRUE)

Method drop()

Drop (delete) UDF

Usage

udf$drop(conn, cascade = FALSE)

Arguments

conn

A DBI::DBIConnection object, as returned by DBI::dbConnect()

cascade

Whether or not to dautomatically rop objects that depend on the UDF (default: FALSE)

Method test()

Run tests TODO

Usage

udf$test(conn)

Arguments

conn

A DBI::DBIConnection object, as returned by DBI::dbConnect()

Method clone()

The objects of this class are cloneable with this method.

Usage

udf$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.